Move Hydrator to FieldCollection namespace.
