+++ config/class.yaml
@@ -21,5 +21,5 @@ services:
   #
   # Hydrators
   #
-  Pimcore\Bundle\StudioBackendBundle\Class\Hydrator\LayoutDefinitionHydratorInterface:
-    class: Pimcore\Bundle\StudioBackendBundle\Class\Hydrator\LayoutDefinitionHydrator
+  Pimcore\Bundle\StudioBackendBundle\Class\Hydrator\FieldCollection\LayoutDefinitionHydratorInterface:
+    class: Pimcore\Bundle\StudioBackendBundle\Class\Hydrator\FieldCollection\LayoutDefinitionHydrator
